Many Eyes - Yet Another Data Collector
Many Eyes is a data collector that can be used for threat hunting and
data collection. Excuse the bad code, it is quite hacky and is provided
only for demo purposes.

It is modular, multithreaded and basic, created
to teach interns and analysts on utilizing python for pentesting and
data collection.

Public release has shodan & twitter support, personal one has 5
(virustotal, censys and pastebin) platforms for now.